Add statusBarColor to navigator props docs
Reading the docs I couldn't see a way to simply change the color of the status bar on Android. Only via https://github.com/wix/react-native-navigation/issues/571#issuecomment-267255350 I saw there was a prop to do this.

(Assuming this is Android only since iOS only allows for light or dark).