Moved Deep Link doco in line with Screen API
I wasnt entirely sure how to change the link referenced in `handleDeepLink` to navigate directly to `deep-links.md`, so I thought Id move it back into here where Im hoping it was intended. Happy for feedback for the correct way to reference an external MD file if that's the preference.