Add ability to reset navigation in the same way as you start an app
The method startSingleScreenApp will now return the freshly created
navigatorID. Using that together with updateSingleScreen app you can now
reset the navigation from outside a screen.

This was implemented to be able to listen to navigation deeplinks sent
from a drawer, and flip the page without reimplementing the same logic
on every drawer page.