Fix order of `registerComponentWithRedux` arguments
The actual `registerComponentWithRedux` function definition is the following:
```
    registerComponentWithRedux(componentName, getComponentClassFunc, ReduxProvider, reduxStore) {
        return this.componentRegistry.registerComponent(componentName, getComponentClassFunc, ReduxProvider, reduxStore);
    }
```

Note the order of `store` and `provider` are the opposite as shown in the original documentation.  This will fix it, but it's worth mentioning that v1 has the order shown in the original documentation, so it would be nice to keep the parameter order the same in v2.