Refactor ComponentWrapper
ComponentWrapper.wrap was a static method which handled both regular components and components wrapped with Redux provider.
This change is needed to support React’s Context api. Context api shares similarities with redux, as components are also wrapped with a Provider.
The only difference is that the component can be wrapped with multiple providers.