Document options for Peek and Pop previews
This is the Information I needed to figure out how to implement Peek and Pop previews. Since the rest of the documentation is fairly concise I was wondering wether it would be appropriate to include "recipe" type things, like disabling scroll during preview, reacting to preview state in the previewed screen etc here or maybe in some other section?