Always set prefersLargeTitles to YES on iOS 11+
If the value of `navigationBar.prefersLargeTitles` is changed while
using largeTitles, there will be a weird navigation animation when going
from a screen that uses largeTitles to a screen that doesn't use
largeTitles.

Instead of turning on/off the largeTitles with prefersLargeTitles, I
found it's better to use the set the `navigationItem.largeTitleDisplayMode`
in `UIViewController+RNNOptions.m`. Switching this value between
`UINavigationItemLargeTitleDisplayModeAlways` and
`UINavigationItemLargeTitleDisplayModeNever` will allow dynamic setting
of the largeTitles and not cause the animation problems.