Removes jsLocation and RCTBridgeDelegate code and changes bootstrap signatures
- Breaking change:
The reason behind that is, there's no pros in handling js location in react-native-navigation
(this feels outside the scope of that library) and there are various cons of doing so.
For starters, it forces us to use BridgeManager as the responder of RCTBridgeDelegate, which
used to be fine. Now with TurboModules though there's a need of responding to other protocols,
such as the RCTCxxBridgeDelegate, and the RCTTurboModuleManagerDelegate all of which are declared
via the RCTBridge initialization (which happened and still happens in RNNBridgeManager)

Changing the responder of RCTBridgeDelegate from RNNBridgeManager to the actual AppDelegate of
the app being built, allows the end user to integrate with TurboModules and customize their code as much as they
want, plus it provides a future proof  way of making sure RNN doesn't break react-native additions
especially since the turbo modules api is fairly new and likely to change.