Include componentId in props passed to static options generator
Up until now the static options generator was invoked with passProps while the component's props contain passProps as well as componentId. This commit eliminates this inconsistency as it's a source for confusion and from now on passProps will also contain the componentId.

Note that if the user has specified a prop named `componentId` then Navigation won't override it.