Use correct ScaleType when animating same element multiple times
When animating images in shared element transition, we change their ScaleType to an InterpolatingScaleType.
At the end of the animation we need to restore the ScaleType back to its original value but for some reason this doesn't seem to work.

The current solution recursively find the original scale type used for the destination image.

Related to #6505