Fix title and subtitle font styling issue with mergeOptions
When a screen is pushed into a stack without a title/subtitle and the title/subtitle are then updated later on with a call to mergeOptions - font styles like color, size and typeface were not applied.

This commit ensures that resolved font styles are always applied when updating title/subtitle with mergeOptions.
Fixes #6597