Set default title and subtitle font sizes in style xml
The title and subtitle's initial font sizes are computed in SP instead of DP. This caused an issue when screens were pushed without title/subtitle and later on updated title/subtitle via mergeOptions.
The problem was that the new texts were still using the default SP font size which is slightly different than the DP size.

This commit sets default font sizes in dp.