Fix ReactImageView sporadic flicker in pop during SET
When animating shared elements in pop, the appearing screen needs a bit of time to be settle in and for the React images to be displayed.
Seems like a good old post does the job in this case. This optimisation is done only for ReactImages. For FastImage we continue checking thier measured bounds as before.