Fix border radius getting clipped in Image elements
While animating ReactImageView in shared element transition, we're changing the bounds of the destination image to max(sourceImage.bounds, destinationImage.bounds) and apply a clipping rect.
This apparently messes up border radius as Fresco isn't aware of the clipping rect and draws the corners according the actual image bounds.

This commit splits the clip bounds animator into two animators, one for Image and another for FastImage.
In the ReactImageView animator we update both clip bounds and the drawable bounds