Check if the drawable has been scaled to fit before considering it measured
Seems like native images load the drawable in two steps. In the second step the image is scaled to fit its bounds. If we return the view before the image has been scaled properly, the shared element transition will use incorrect values when calculating dimensions. This will cause the image not to transition correctly into place.

This change is related to native ImageViews and does not affect ReactImageView.