fix(iOS): avoid cutting-off text on bottom bar buttons on iPad.
For some reason, this change resolves the issue in the attached snapshots.

When `centered=YES`, setting the text alignment to `NSTextAlignmentCenter`
may incidentally cause the text layout engine to more effectively manage the
available space. Doesn't seem like it affects anything in iPhones.

Overall, it seems like a bug in UIKit..